Refactor search and extraction tools
